web と CLI で共用するルーチンを dropitemseditor.py に切り離す。
